one finufft.h, reorganization of header inclusion
