+++ src/common.cpp
@@ -3,6 +3,7 @@
 
 #include "common.h"
 #include "defs.h"
+#include <utils.h>
 #include <fftw3.h>
 #include <math.h>
 #include <stdio.h>
